User Refactor

+++ src/main/java/com/zeroway/user/service/UserService.java
@@ -18,7 +18,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
 import java.io.IOException;
-import java.util.List;
 import java.util.Optional;
 
 import static com.zeroway.common.BaseResponseStatus.*;
@@ -65,12 +64,14 @@ public class UserService {
 
         // 레벨 1
         Optional<Level> levelOptional = levelRepository.findById(1);
-        user.setLevel(levelOptional.get());
+        if (levelOptional.isPresent()) {
+            user.changeLevel(levelOptional.get());
+        }
 
         try {
             if (profileImg != null && !profileImg.isEmpty()) {
                 String userProfileUrl = s3Uploader.uploadFile(profileImg, "userProfile");
-                user.setProfileImgUrl(userProfileUrl);
+                user.uploadProfileImg(userProfileUrl);
             }
         } catch (IOException e) {
             throw new BaseException(FILE_UPLOAD_ERROR);
@@ -83,7 +84,7 @@ public class UserService {
     /**
      * 토큰 두종류 발급
      */
-    private PostUserRes postUser(User user) throws BaseException {
+    private PostUserRes postUser(User user) {
         String refreshJwt = jwtService.createRefreshToken(user.getId());
         String accessJwt = jwtService.createAccessToken(user.getId());
 
@@ -125,7 +126,11 @@ public class UserService {
         try {
             // 토큰 만료 확인
             Long userIdx = jwtService.getUserIdx();
-            User user = userRepository.findById(userIdx).get();
+            User user = null;
+            Optional<User> optionalUser = userRepository.findById(userIdx);
+            if (optionalUser.isPresent()) {
+                user = optionalUser.get();
+            }
             user.setStatus(StatusType.LOGOUT);
             jwtService.deleteRefreshToken(userIdx);
         } catch (BaseException e) {
@@ -137,37 +142,47 @@ public class UserService {
      * 회원탈퇴
      */
     public User signout() throws BaseException {
+        User user = null;
+
         try {
             Long userIdx = jwtService.getUserIdx();
-            User user = userRepository.findById(userIdx).get();
+            Optional<User> optionalUser = userRepository.findById(userIdx);
+            Optional<Level> levelOptional = levelRepository.findById(1);
+            if (optionalUser.isPresent() && levelOptional.isPresent()) {
+                user = optionalUser.get();
 
-            user.signout("알 수 없음", "email@gmail.com", null, levelRepository.findById(1).get(), StatusType.INACTIVE);
-            jwtService.deleteRefreshToken(userIdx);
-            return user;
+                user.signout("알 수 없음", "email@gmail.com", null, levelOptional.get(), StatusType.INACTIVE);
+                jwtService.deleteRefreshToken(userIdx);
+            }
         } catch (BaseException e) {
             throw new BaseException(DATABASE_ERROR);
         }
+
+        return user;
     }
 
     /**
      * 회원정보 수정
      */
     public void patchUser(MultipartFile profileImg, PatchUserInfo patchUserInfo) throws BaseException {
-        User user = null;
         try {
             Long userIdx = jwtService.getUserIdx();
-            user = userRepository.findById(userIdx).get();
+            User user = null;
+            Optional<User> optionalUser = userRepository.findById(userIdx);
+            if (optionalUser.isPresent()) {
+                user = optionalUser.get();
+            }
 
             if (patchUserInfo != null) {
                 String nickname = patchUserInfo.getNickname();
-                user.setNickname(nickname);
+                user.changeNickname(nickname);
             }
 
             if (profileImg == null) {
-                user.setProfileImgUrl(null);
+                user.uploadProfileImg(null);
             } else if (!profileImg.isEmpty()) {
                 String userProfile = s3Uploader.uploadFile(profileImg, "userProfile");
-                user.setProfileImgUrl(userProfile);
+                user.uploadProfileImg(userProfile);
             }
         } catch (IOException e) {
             throw new BaseException(FILE_UPLOAD_ERROR);
